Pushing the sign-up for updates additions
