+++ public/signup.html
@@ -1,0 +1,55 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Get Updates</title>
+    <link href="../../public/images/favicon.png">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
+    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+    <link rel="preconnect" href="https://fonts.gstatic.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com">
+    <link href="https://fonts.googleapis.com/css2?family=Salsa&display=swap" rel="stylesheet">
+    <link rel="shortcut icon" type="image/x-icon" href="./images/favicon.png">
+    <script type="text/javascript" src="http://cdnjs.cloudflare.com/ajax/libs/jquery/2.1.1/jquery.min.js"></script>
+    <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/2.0.0/handlebars.min.js" ></script>
+    <script src="css/style.css"></script>
+</head>
+<body>
+    
+    <script type="text/javascript" src="http://cdnjs.cloudflare.com/ajax/libs/jquery/2.1.1/jquery.min.js"></script>
+
+    <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/2.0.0/handlebars.min.js" ></script>
+
+
+    <div class="container-fluid col s12 m8 l9" id="applyForm">
+
+        <h2>Join our network!</h2>
+        <p>Enter your name and email to receive weekly updates about our eligible pets!</p>
+        <div class="row">
+            <form class="col m8 offset-m2 s12">
+                <div class="row">
+                    <div class="input-field col s12">
+                        <input id="Name" type="text" class="validate">
+                        <label for="news-name">Full Name</label>
+                    </div>
+                <div class="row">
+                    <div class="input-field col s12">
+                        <input id="Email" type="text" class="validate">
+                        <label for="email">Email</label>
+                    </div>
+                </div>
+                <button class="btn waves-effect waves-light #37474f blue-grey darken-3" type="submit" name="action"
+                    id="verifyBtn">Send me pet
+                    updates!
+                    <i class="material-icons right">account_circle</i>
+                </button>
+        </div>
+    </div>
+</div>
+</form>
+</body>
+
+<script src="./js/signup.js"></script>
+
+</html>
