+++ views/partials/navbar-partial.handlebars
@@ -3,7 +3,7 @@
         <a href="#" class="brand-logo">Peer 2 Paws</a>
         <ul id="nav-mobile" class="right hide-on-med-and-down">
             <li><a href="/">Home</a></li>
-            <li><a href="sign-in">Get Updates/Sign-up</a></li>
+            <li><a href="signup">Get Updates/Sign-up</a></li>
             <li><a href="contact">Contact Us</a></li>
         </ul>
     </div>
