--- views/sign-in.handlebars
@@ -1,39 +1,0 @@
-{{> navbar-partial}}
-
-<div class='row'>
-
-    {{> sidebar-partial}}
-    <div class="container-fluid col s12 m8 l9" id="applyForm">
-        {{!-- Show sign-in info here --}}
-        <h2>Join our network!</h2>
-        <p>Enter your name and email to receive weekly updates about our eligible pets!</p>
-        <div class="row">
-            <form class="col m8 offset-m2 s12">
-                <div class="row">
-                    <div class="input-field col s12">
-                        <input id="FirstName" type="text" class="validate">
-                        <label for="news-first-name">First Name</label>
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="input-field col s12">
-                        <input id="news-last-name" type="text" class="validate">
-                        <label for="news-last-name">Last Name</label>
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="input-field col s12">
-                        <input id="Email" type="text" class="validate">
-                        <label for="email">Email</label>
-                    </div>
-                </div>
-                <button class="btn waves-effect waves-light #37474f blue-grey darken-3" type="submit" name="action"
-                    id="verifyBtn">Send me pet
-                    updates!
-                    <i class="material-icons right">account_circle</i>
-                </button>
-                <hr>
-
-        </div>
-    </div>
-</div>
