+++ views/specificType.handlebars
@@ -1,0 +1,27 @@
+{{> navbar-partial}}
+
+
+<div class="row">
+
+    {{> sidebar-partial}}
+
+    <div class="container-fluid col s12 m8 l9">
+
+        <h2>All Our</h2>
+
+        <ul class="col s7 push-s1">
+            {{#each pets}}
+            <p id="petsCard">
+                Type: {{TypeName}}<br>
+                Name: {{AnimalName}}<br>
+                Age: {{Age}}<br>
+                About: {{About}}<br>
+                Owner: {{OwnerName}}<br>
+                Email: {{OwnerEmail}}<br>
+            </p>
+            {{/each}}
+        </ul>
+
+    </div>
+
+</div>
